Work around problem with tm:::TextDocumentMeta
When constructing a PlainTextDocument, tm does special handling for its 'standard' metadata fields: author, datetimestamp, description, heading, id, language, and origin.

Unfortunately, it does so using "if(!is.null(meta$origin))" etc., with the $ notation running the risk of partial matching against other elements when the original field is absent. (In particular, where there is an 'originalurl' field in the RISJbot data, but no 'origin' found). There shouldn't be any other clashes in standard RISJbot output.

This patch works around this potential problem.